% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQ-LIC-TN.R
\name{BQ_LIC_TN}
\alias{BQ_LIC_TN}
\title{Bayesian inference of quantile regression models with linear inequality constraints based on truncated Laplace prior}
\usage{
BQ_LIC_TN(x, y, p, R, b)
}
\arguments{
\item{x}{data X, n*d matrix}

\item{y}{data Y, n*1 vector}

\item{p}{quantile level}

\item{R}{constraint matrix with the dimension m*d}

\item{b}{constraint vector with the dimension m}
}
\value{
estimated parameter, 1*d vector
}
\description{
Bayesian inference of quantile regression models with linear inequality constraints based on truncated Laplace prior
}
\examples{
n <- 200
d <- 3
beta <- c(0.3,0.7,-0.5)
R <- matrix(c(1,0,0,1,0,1),2,3)
b <- c(0.2,0)
p <- 0.5
x <- rmvnorm(n,rep(0,d),diag(d))
epsilon <- rnorm(n,0,1)
epsilon1 <- epsilon-qnorm(p)
y <- x\%*\%beta+epsilon1
intsp <- c(0.5,0.5,-0.1) #initial value
beta_LIC_TN <- BQ_LIC_TN(x,y,p,R,b)
}
